refactor: move breadcrumbs to shared
