+++ src/components/pages/docs/post-layout/post-layout.tsx
@@ -1,12 +1,14 @@
 import Link from '@/components/shared/link';
 
-import { Breadcrumb, TableOfContents as TOCProps } from '@/types/docs';
+import { TableOfContents as TOCProps } from '@/types/docs';
+
+import { Breadcrumb } from '@/types/breadcrumb';
 
 import ExternalLinkIcon from '@/svgs/external.inline.svg';
 
-import Breadcrumbs from '../breadcrumbs';
 import Navigation, { type NavigationProps } from '../navigation';
 import TableOfContents from '../table-of-contents';
+import Breadcrumbs from '@/components/shared/breadcrumbs';
 
 const FILE_ORIGIN_PATH = 'https://github.com/bytebase/bytebase.com/tree/main/content/docs';
 
@@ -28,7 +30,7 @@ const PostLayout = ({
   return (
     <>
       <article className="col-span-6 col-start-4 flex flex-col lg:col-span-9 md:col-span-full">
-        {breadcrumbs.length > 0 && <Breadcrumbs breadcrumbs={breadcrumbs} />}
+        {breadcrumbs.length > 0 && <Breadcrumbs className="mb-7 mt-[13px]" items={breadcrumbs} />}
         <h1 className="mt-2.5 text-44 font-bold leading-extra-tight tracking-tighter text-gray-15 2xl:mt-1.5 lg:text-36 md:text-32 sm:mt-0 sm:text-30">
           {title}
         </h1>
