+++ src/types/breadcrumb.ts
@@ -1,0 +1,4 @@
+export type Breadcrumb = {
+  title: string;
+  url?: string;
+};
